Fix: Fixing light & shadows artifacts

+++ Texture.cpp
@@ -33,7 +33,7 @@ Texture::Texture(int Width, int Height, TextureType loadType) :
    
     if (loadType == TextureType::FRAMEBUFFER_ATTACHMENT)
     {
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
+        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, NULL);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
         glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
